gheatmap: add error message for badly formatted @data
gheatmap requires that the heat map data contain named rows that match
the tree species. If the dataframe is incorrectly formatted, the user
will later receive the error "must request at least one colour from a
hue palette", which is confusing. Detect this error and report it when
it is found.

Signed-off-by: Douglas Fuller <douglas.fuller@gmail.com>